+++ src/components/Card.tsx
@@ -1,4 +1,4 @@
-import React, { MouseEvent, useState } from 'react';
+import React, { MouseEvent, ReactNode, useState } from 'react';
 import { ICampaignItem } from '../types/campaign';
 import styled from 'styled-components';
 import {
@@ -7,16 +7,11 @@ import {
   CardActions,
   CardContent,
   CardHeader,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableRow,
   Popover,
   Typography,
 } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import { format } from 'date-fns';
+import CardTable from './CardTable';
 
 interface CardProps {
   campaignItem: ICampaignItem;
@@ -25,7 +20,6 @@ interface CardProps {
 
 function Card({ campaignItem, onDelete }: CardProps) {
   const [anchorEl, setAnchorEl] = useState<HTMLButtonElement | null>(null);
-  const date = new Date(campaignItem.startDate);
   const open = Boolean(anchorEl);
 
   const navigate = useNavigate();
@@ -50,40 +44,7 @@ function Card({ campaignItem, onDelete }: CardProps) {
     <BasicCard variant='outlined' style={{ borderRadius: '12px' }}>
       <CardHeader title={campaignItem.title} sx={{ pb: 0 }} />
       <CardContent>
-        <TableContainer>
-          <Table size='small'>
-            <TableBody>
-              <Row style={{ borderTop: '1px solid rgba(224, 224, 224, 1)' }}>
-                <TableCell>상태</TableCell>
-                <TableCell>{campaignItem.status === 'active' ? '진행중' : '종료'}</TableCell>
-              </Row>
-              <Row>
-                <TableCell>광고 생성일</TableCell>
-                <TableCell>{format(date, 'yyyy-MM-dd')}</TableCell>
-              </Row>
-              <Row>
-                <TableCell>일 희망 예산</TableCell>
-                <TableCell>{campaignItem.budget / 10000}만원</TableCell>
-              </Row>
-              <Row>
-                <TableCell>광고 수익률</TableCell>
-                <TableCell>{campaignItem.report.roas}%</TableCell>
-              </Row>
-              <Row>
-                <TableCell>매출</TableCell>
-                <TableCell>
-                  {Math.floor(campaignItem.report.convValue / 10000).toLocaleString()}만원
-                </TableCell>
-              </Row>
-              <Row>
-                <TableCell>광고 비용</TableCell>
-                <TableCell>
-                  {Math.floor(campaignItem.report.cost / 10000).toLocaleString()}만원
-                </TableCell>
-              </Row>
-            </TableBody>
-          </Table>
-        </TableContainer>
+        <CardTable campaignItem={campaignItem} />
       </CardContent>
       <ButtonWrapper sx={{ p: 0, pb: 2 }}>
         <Button size='small' variant='outlined' color='inherit' onClick={handleModifyClick}>
@@ -138,17 +99,6 @@ const BasicCard = styled(DefaultCard)`
   /* desktop */
   @media (min-width: 1024px) {
     width: calc(33.33333% - 24px);
-  }
-`;
-
-const Row = styled(TableRow)`
-  & > td:first-child {
-    padding-left: 0;
-    color: gray;
-  }
-
-  & > td {
-    padding: 0.5rem;
   }
 `;
